AWS Lambda Compatibility
When attempting to use caesium in AWS Lambda, this error occurs:

```
{
      "errorMessage": "Error loading class app.handler:
    caesium/binding$✨",
      "errorType": "java.lang.NoClassDefFoundError"
}
```

There appears to be an issue in Lambda with executing Clojure code that
contains Unicode characters. This error does not occur in an EC2
instance using the same AMI that Lambda uses.

So while it is not wrong that caesium had named this macro as a Unicode
character, and this is surely Lambda's bug that they should fix, in the
interim this changes caesium to use a name that is exactly as
descriptive (✨ vs `magic-sparkles`), but is capable of being executed by
AWS Lambda.

I chose the name `magic-sparkles` because the test was already called
`magic-sparkles-test`, which indicated that this was how the Unicode
character was pronounced.