Add three aead constructions:
- ChaCha20Poly1305 - The original ChaCha20-Poly1305 construction can safely
encrypt a pratically unlimited number of messages with the same key,
without any practical limit to the size of a message (up to ~ 2^64 bytes).

- ChaCha20Poly1305-Ietf - Variant of the ChaCha20-Poly1305 construction
can safely encrypt a pratically unlimited number of messages, but
individual messages cannot exceed 64*(2^32)-64 bytes (approximatively 256 GB).

- XChaCha20Poly1305 - The XChaCha20-Poly1305 construction can safely
encrypt a practically unlimited number of messages with the same key,
without any practical limit to the size of a message (up to ~ 2^64 bytes).

As an alternative to counters, its large nonce size (192-bit) allows
random nonces to be safely used.