Fix the `:deprecations` warning on `s/generate-signing-keys`
The fn is marked as deprecated and eastwood sincerely complains on
it. Yet we want to test the non-deprecated alias is properly
provided and thus need to use the deprecated symbol in a test.

This adds an eastwood config file `eastwood.clj` in the project
root to stop eastwood from complaining on the project's deprecated
symbols—just `caesium.crypto.sign/generate-signing-keys` for now.